combine plotting scripts
